tg_debug_bot_v1TатАФ ¦¦¦-¦¬¦¬TП/requirements.txt
anyio==4.0.0
APScheduler==3.6.3
asgiref==3.7.2
cachetools==4.2.2
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2.7
exceptiongroup==1.1.3
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
python-telegram-bot==13.15
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tornado==6.1
typing_extensions==4.8.0
tzlocal==5.2
urllib3==2.1.0
